# handler/auth/password/sign_up.go
package handler

import (
	"encoding/json"
	"log"
	"net/http"

	"github.com/ecofriends/authentication-backend/authentication"
	"github.com/ecofriends/authentication-backend/model"
	"github.com/ecofriends/authentication-backend/service"
	"github.com/ecofriends/authentication-backend/util"
	validators "github.com/ecofriends/authentication-backend/validator"
	"github.com/google/uuid"
)

// SignUp handles user registration
// @Summary Register a new user
// @Description Create a new user account
// @Tags authentication
// @Accept json
// @Produce json
// @Param request body util.SignUpRequestBody true "Sign up credentials"
// @Success 200 {object} util.Response
// @Failure 400 {object} util.Response
// @Failure 500 {object} util.Response
// @Router /auth/sign-up [post]
func SignUp(dbService *service.DatabaseProvider, w http.ResponseWriter, r *http.Request) {
	// Store the request body
	var body = util.SignUpRequestBody{}

	// Read response body into body struct
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		msg := "Bad request, username, email or password not present"
		util.JsonResponse(w, msg, http.StatusBadRequest, nil)
		return
	}

	// Sanitize user input
	util.SanitizeUserInput(&body)
	log.Printf("[LOG]: Sanitized user input")

	// Validate the user input
	if err := validators.ValidateUserInput(&body); err != nil {
		msg := util.CapitalizeFirstLetter(err.Error())
		util.JsonResponse(w, msg, http.StatusBadRequest, nil)
		return
	}

	// Do not create a new user if the user already exists
	userExists, err := dbService.Repo.UserExists(r.Context(), body.Email, body.Username)
	if err != nil {
		log.Println(err)
		msg := "Could not check if user already exists"
		util.JsonResponse(w, msg, http.StatusInternalServerError, nil)
		return
	}

	// Respond with a bad request if the user exists
	if userExists {
		msg := "A user with those credentials already exists"
		util.JsonResponse(w, msg, http.StatusBadRequest, nil)
		return
	}

	// Prepare the user data for insertion
	var user = model.User{
		ID:       uuid.New(),
		Username: body.Username,
		Email:    body.Email,
		Password: body.Password,
	}

	// Generate a JSON Web token that can be sent to the user
	token, err := authentication.CreateJWToken(user.ID)
	if err != nil {
		log.Println(err)
		msg := "Failed to create token"
		util.JsonResponse(w, msg, http.StatusInternalServerError, nil)
		return
	}

	// Insert the user into the database
	err = dbService.Repo.InsertUser(r.Context(), user)
	if err != nil {
		log.Println(err)
		msg := "Could not insert user into database"
		util.JsonResponse(w, msg, http.StatusInternalServerError, nil)
		return
	}

	// Create the user payload
	var userPayload = util.UserPayload{
		ID:       user.ID,
		Username: user.Username,
		Email:    user.Email,
	}

	// Set the token cookie and send the response
	cookie := util.CreateTokenCookie(token)
	http.SetCookie(w, &cookie)
	util.JsonResponse(w, "Successfully inserted user into database", http.StatusOK, userPayload)
}